junit test rev2
